Think I figured out how I will do this ...
I'll modify the configuration file (itself needs more info) to have a view mode. Depending on the view that's how we'll do stuff such as set the poses, e.g. for the closer or standard view.

Good news is that I tried the image viewer, and was able to tell the HSR to grab and then pull to the edge. So that's good. Didn't try moving to other side but that should be trivial.

Will need to modify table_top and probably best to make more poses, don't overwrite what we have now.

Tomorrow: focus on getting data collection done the way I was doing it, and test both the old view and the newer view. Then that should give us a lot of information ...

OH and in addition to that (if possible) try to load in a trained neural network from today.