Improved data logging and frames
now going to the top and back, the hsr is facing towards the direction of motion, not sideways

i saved a bunch of timing information

the deployment script has a phase argument so I don't need to comment out that stupid rospy.spin thingy

i saved one rollout from start to finish. It seemed to go well. Next up, need to evaluate this, so I saved as a dict. Will now do that offline, then collect more demos in evening